% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmcp-stats-package.R
\docType{package}
\name{rmcp.stats-package}
\alias{rmcp.stats}
\alias{rmcp.stats-package}
\title{RMCP statistical utilities}
\description{
Shared functions that standardize JSON input, output, progress, and
validation for RMCP's R scripts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/finite-sample/rmcp}
  \item Report bugs at \url{https://github.com/finite-sample/rmcp/issues}
}

}
\author{
\strong{Maintainer}: RMCP Team \email{contact@gsood.com}

Authors:
\itemize{
  \item RMCP Team \email{contact@gsood.com}
}

}
\keyword{internal}
